Day 1 done.
Took me a while to work out some basic stuff with ocaml, but writing the code itself was
quite nice. The types in the editor are fantastic and made it easy to spot errors quickly.
It also had the nice property of once it compiled it pretty much worked.

The main frustration was with building and running the code. There is a lot of different advice
out there depending on if you are using dune/oasis/ocamlbuild etc.

At the moment I am compiling with:

`ocamlbuild part1.native -pkgs Str`

And then running with:

`./part1.native`

and that seems to work okay for now, but it's not clear what the 'standard' is in the ocaml community.
Docs where pretty good, which I think is much easier with a statically typed language. Especially languages
like this where the type information is most of what you need. Looking forward to moving forward and learning
some more about ocaml as part of this years AoC.