Day 2 done!
I made one small mistake with data structures at one point (doing rev and hd in the wrong order)
but overall found this day much easier now that I am getting used to the language on the whole and
the tooling around it.

I am nervous that the problems may get less conducive to this sort of programming though. At the moment
parsing and processing is easy with functional control strucutres, but it may get harder if we need
to index into arrays or anything. Guess we will see?

For now though really enjoying this and found the main problem trying to work out Rock paper scissors!
I wasted a long time debugging an issue that turned out to be just incorrect transcription of the game results!
So that was kind of annoying, but I guess I should've made a data structure with better names :shrug:
Still need to learn some more about that stuff with Ocaml.