Day 3 done!
This one was a little trickier with Ocaml I found. I knew how to do this puzzle
imperatively, in python I probably would've used a Dict or Set to find common elements
doing it with Ocaml I had to get a little creative and wasn't super happy with my solution
in the end.

There are sets, hashtables and assoc lists in ocaml that potentially could've helped me out
with this one. But in the end I just got a little creative with filtering and mapping to get
the data in an easy to work with format.

Hopefully as I get more experience with Ocaml thinking this way will be easier and I will also
probably explore the standard library some more and make use of more data structures. Anyway,
onwards and upwards I guess! :shrug: