Day 5 done!
Today was pretty fun! I got to try out the stack data structure which was pretty cool.
I did get caught out at one point when the `int_of_char` function returned ascii code
instead of parsing the int. I also struggled a bit with working out how to intialise
the list of stacks. I probably didn't do this correctly in the end, but hey it worked.

This felt pretty fun as the problem mapped well to functional programming. It was for
the most part a parsing task and then a matter of applying instructions to the data
structures. This was pretty easy to write compare to other days, but I don't know whether
that's because I'm getting better at Ocaml or whether this problem was just a bit nicer
for it.

I also used Ocaml a little more imperatively today, it has the ability to be OO style
as well as more functional, so maybe I should try explore that more in future days.

My main take-away so far though is that it just feels REALLY nice to write. Defining
and applying functions is seamless and the types in the editor really help to keep things
in my head better over time.