Day 6 done!
Look not super proud of this code, it's pretty bad. But oh well, it did solve the problem pretty quickly.
I might revisit this one later today or tomorrow to make it better, as I wouldn't be surprised if this
'packet parsing' type of problem came up again. It feels like it wasn't fully explored today.

Was pretty happy with how fast I got part1 done, but then spent way too much time mucking around with
lists in part 2. :shrug:

Happy to not have to spend so much time on this day though!