Day 7 done!
Struggled a little bit with some dumb bugs today. I had to take a break and go for a bike
ride, but as I was riding I realised that the answer was to take my tested function (strip_trailing_dir)
that I was happy with how it worked and use it recursively to get all the parents.

This was quite nice, and gels super well with the ocaml way of doing things, so the code was much
nicer this way too.

Was a little frustrated with some of the issues, especially after I think I did a really good job of laying out
in comments the steps to solve the problem. I probaly got 90% of the way there quite quickly, but then made lots of little
bugs toward the end that slowed me down. In future I should reach for recursion quicker in situations like this I think.

All in all very cool problem though. Hopefully be back tomorrow!