Day 8 done!
I was pretty happy with today. I got part1 done in a reasonable-ish amount of time and the code
wasn't too bad to implement. I ran out of time at the end of the day for part 2 though, as I was busy in the evening!

Once I jumped back on in the morning it didn't take too long to solve, but I did get caught on a couple
of little things. Usually stemming from my misunderstanding of the stdlib. (I got the argument order to List.append the
wrong way round)

This will improve with time, and I think I am slowly improving writing nicer code in Ocaml and seeing solutions that can
be more easily mapped to the way ocaml works.