Day 10 done!
Did some dumb stuff with day 9 so I will probably come back to that, but needed a break from it.
So I did day 10! Pretty cool problem today, I quite enjoyed it, and was quite happy with how part 2
turned out. The idea to convert addx instructions to noop, addx worked really well for simplifying the
modelling of the computer execution.

Also used custom record types for the first time with ocaml and found it quite nice for simplifying
the readability of the code. Hopefully I'll continue to learn more about the language as we go through the
contest, but so far I'm quite enjoying ocaml.