Day 11 done!
This was a really cool day! I enjoyed the problem a lot and was pretty happy with my solution.

As part of tackling this problem I learnt more about Ocaml's record type and how you can create
mutable fields on records. This made the code so much easier to read and I was happy with how it
came together.

I did get a bit stuck on part 2 for a bit with the large numbers. I modified it to remove the '/ 3'
and then set the rounds to '10000' but the answer was way too low. This made me suspicious that an
overflow was occuring so I tried out the Big_int library. Running this caused an error warning of
overflowing the big_int library. So at this stage I realised there must be some maths we can use to simplify
things. Looking at the problem again it had a hint that we needed to 'manage' the worry some how. I
realised that you always test for divisibility so using mod seemed a good idea. I just needed to find a divisor
which wouldn't screw with those tests. The easiest way I could think of was to multiply the divisors together.

It was cool going this rather scenic route to the solution, as I learnt more about using opam and installing
3rd party libaries. The docs on this aren't great, so it was a good thing to learn even though I didn't need it
in the end to solve the problem.