Day 9 done!
Finally solved the pesky problem with this task. I missed some directions from my diagonal code which
caused problems with the larger input but not with the example inputs so it was frustrating to solve.

This did give me some good experience with debugging in ocaml, which can be kind of tricky when doing
recursion/mapping and stuff. Overall it was an okay experience, and though I think my overall approach
was sound, I was a little negligent of the details which came back to bite me. It's a good lesson for future.

I'm also not super wrapped with the code in the end, the match statements have way too many arms, but I wanted
to catch lots of different code paths while debugging so they are left there for now. I may come back and tidy this up,
who knows?