Day 14 done!
After a couple of days break, I'm back! Will try come back to the other puzzles when I get some
more time, but happy to have got this one done. I struggled a little with the while loop, the conditional
seemed like it wasn't evaluating sometimes, but idk, managed to get everything working in the end and was
relatively happy with how natural this felt to write.